+++ dist_node/inline/expand.js
@@ -1,8 +1,7 @@
 /*
- * THIS FILE IS AUTO GENERATED FROM 'lib/inline/expand.kep'
+ * THIS FILE IS AUTO GENERATED from 'lib/inline/expand.kep'
  * DO NOT EDIT
-*/
-"use strict";
+*/"use strict";
 var ast_declaration = require("khepri-ast")["declaration"],
     ast_expression = require("khepri-ast")["expression"],
     ast_pattern = require("khepri-ast")["pattern"],
@@ -16,18 +15,25 @@ var ast_declaration = require("khepri-ast")["declaration"],
     rename = __o1["rename"],
     __o2 = require("../builtin"),
     builtins = __o2["builtins"],
-    expandCallee, expandCurry, getParmeterIds = map.bind(null, (function(x) {
-        return getUid(x.id);
-    }));
+    expandCallee, expandCurry;
 (expandCallee = (function(uid, callee, args) {
     var target = ((callee.type === "LetExpression") ? callee.body : callee),
-        ids = getParmeterIds(target.params.elements),
+        params = target.params,
+        ids = concat(map((function(x) {
+            return getUid(x.id);
+        }), params.elements), (params.id ? getUid(params.id.id) : [])),
         parameters = target.params,
         bindings = map((function(x, i) {
-            return ast_declaration.Binding.create(null, rename(uid, ids, x), (args[i] || builtins.undefined));
-        }), parameters.elements);
-    return ast_expression.LetExpression.create(null, concat((callee.bindings || []), bindings), rename(uid, ids,
-        target.body));
+            return ast_declaration.Binding.create(null, rename(uid, [getUid(x.id)], x), (args[i] ||
+                builtins.undefined));
+        }), parameters.elements),
+        arg, argBinding = (target.params.id ? ((arg = target.params.id), ast_declaration.Binding.create(null,
+            rename(uid, [getUid(arg.id)], arg), ast_expression.ArrayExpression.create(null, args.map((
+                function(x, i) {
+                    return (bindings[i] ? bindings[i].pattern.id : x);
+                }))))) : []);
+    return ast_expression.LetExpression.create(null, concat((callee.bindings || []), bindings, argBinding),
+        rename(uid, ids, target.body));
 }));
 (expandCurry = (function(uid, base, args) {
     var first, rest, map0, body, target = ((base.type === "LetExpression") ? base.body : base);